BluetoothStandardWeightProfile: reuse common method to decode vendor specific scale user data;
